src/test/java/it/schwarz/jobs/review/coupon/domain/entity/AmountOfMoneyTests.java
package it.schwarz.jobs.review.coupon.domain.entity;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatNullPointerException;
import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

class AmountOfMoneyTests {

    @Test
    void testZero() {
        assertThat(AmountOfMoney.ZERO.toBigDecimal()).isZero();
    }

    @Test
    void testIsGreaterThan() {
        var amount = AmountOfMoney.of("1.23");

        assertTrue(amount.isGreaterThan(AmountOfMoney.of("1.22")));
        assertTrue(amount.isGreaterThan(AmountOfMoney.of("1.00")));
        assertTrue(amount.isGreaterThan(AmountOfMoney.of("0")));
        assertTrue(amount.isGreaterThan(AmountOfMoney.of("-1.23")));
        assertFalse(amount.isGreaterThan(AmountOfMoney.of("1.23")));
        assertFalse(amount.isGreaterThan(AmountOfMoney.of("1.3")));
    }

    @Test
    void testIsLessThan() {
        var amount = AmountOfMoney.of("1.23");

        assertFalse(amount.isLessThan(AmountOfMoney.of("1.22")));
        assertFalse(amount.isLessThan(AmountOfMoney.of("1.00")));
        assertFalse(amount.isLessThan(AmountOfMoney.of("0")));
        assertFalse(amount.isLessThan(AmountOfMoney.of("-1.23")));
        assertFalse(amount.isLessThan(AmountOfMoney.of("1.23")));
        assertTrue(amount.isLessThan(AmountOfMoney.of("1.3")));
    }

    @Test
    void testComparisonWithNull() {
        var amount = AmountOfMoney.of("1.23");
        assertThatNullPointerException().isThrownBy(
                () -> amount.isGreaterThan(null)
        );
    }

    @Test
    void testCreationWithNull() {
        assertThatNullPointerException().isThrownBy(
                () -> AmountOfMoney.of((String)null)
        );
    }
}